move goodstypes urls to applications

+++ applications/views/application_parties.py
@@ -1,4 +1,3 @@
-from django.db import transaction
 from django.http import JsonResponse, HttpResponse
 from rest_framework import status
 from rest_framework.views import APIView
@@ -7,7 +6,7 @@ from applications.enums import ApplicationLicenceType
 from conf.authentication import ExporterAuthentication
 from conf.decorators import only_applications, authorised_users
 from parties.helpers import delete_party_document_if_exists
-from parties.models import UltimateEndUser, ThirdParty, EndUser
+from parties.models import UltimateEndUser, ThirdParty
 from parties.serializers import EndUserSerializer, UltimateEndUserSerializer, ConsigneeSerializer, ThirdPartySerializer
 from users.models import ExporterUser
 
@@ -15,14 +14,14 @@ from users.models import ExporterUser
 class ApplicationEndUser(APIView):
     authentication_classes = (ExporterAuthentication,)
 
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=True)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def post(self, request, application):
         """
         Create an end user and add it to a application
         """
         data = request.data
-        data['organisation'] = str(request.user.organisation.id)
+        data['organisation'] = request.user.organisation.id
 
         serializer = EndUserSerializer(data=data)
         if serializer.is_valid():
@@ -36,33 +35,27 @@ class ApplicationEndUser(APIView):
                 delete_party_document_if_exists(previous_end_user)
                 previous_end_user.delete()
 
-            return JsonResponse(data={'end_user': serializer.data},
-                                status=status.HTTP_201_CREATED)
+            return JsonResponse(data={'end_user': serializer.data}, status=status.HTTP_201_CREATED)
 
-        return JsonResponse(data={'errors': serializer.errors},
-                            status=status.HTTP_400_BAD_REQUEST)
+        return JsonResponse(data={'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
-    @transaction.atomic()
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=True)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def delete(self, request, application):
         """
         Delete an end user and their document from an application
         """
-        try:
-            end_user = EndUser.objects.get(id=application.end_user.id)
-        except UltimateEndUser.DoesNotExist:
-            return JsonResponse(data={'errors': 'request invalid'},
-                                status=400)
+        end_user = application.end_user
+
+        if not end_user:
+            return JsonResponse(data={'errors': 'consignee not found'}, status=status.HTTP_404_NOT_FOUND)
 
         if end_user.organisation != request.user.organisation:
-            return JsonResponse(data={'errors': 'request invalid'},
-                                status=400)
-
-        delete_party_document_if_exists(end_user)
+            return JsonResponse(data={'errors': 'request invalid'}, status=status.HTTP_400_BAD_REQUEST)
 
         application.end_user = None
         application.save()
+        delete_party_document_if_exists(end_user)
         end_user.delete()
 
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
@@ -71,7 +64,7 @@ class ApplicationEndUser(APIView):
 class ApplicationUltimateEndUsers(APIView):
     authentication_classes = (ExporterAuthentication,)
 
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=False)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def get(self, request, application):
         """
@@ -81,39 +74,57 @@ class ApplicationUltimateEndUsers(APIView):
 
         return JsonResponse(data={'ultimate_end_users': ueu_data})
 
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=True)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def post(self, request, application):
         """
         Create an ultimate end user and add it to a application
         """
         data = request.data
-        data['organisation'] = str(request.user.organisation.id)
+        data['organisation'] = request.user.organisation.id
 
         serializer = UltimateEndUserSerializer(data=data)
         if serializer.is_valid():
             ultimate_end_user = serializer.save()
-            application.ultimate_end_users.add(str(ultimate_end_user.id))
-            application.save()
+            application.ultimate_end_users.add(ultimate_end_user.id)
 
-            return JsonResponse(data={'ultimate_end_user': serializer.data},
-                                status=status.HTTP_201_CREATED)
+            return JsonResponse(data={'ultimate_end_user': serializer.data}, status=status.HTTP_201_CREATED)
 
-        return JsonResponse(data={'errors': serializer.errors},
-                            status=status.HTTP_400_BAD_REQUEST)
+        return JsonResponse(data={'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+
+
+class RemoveApplicationUltimateEndUser(APIView):
+    authentication_classes = (ExporterAuthentication,)
+
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
+    @authorised_users(ExporterUser)
+    def delete(self, request, application, ueu_pk):
+        """
+        Delete an ultimate end user and remove it from the application
+        """
+        try:
+            ultimate_end_user = application.ultimate_end_users.get(id=ueu_pk)
+        except UltimateEndUser.DoesNotExist:
+            return JsonResponse(data={'errors': 'ultimate end user not found'}, status=status.HTTP_404_NOT_FOUND)
+
+        application.ultimate_end_users.remove(ultimate_end_user.id)
+        delete_party_document_if_exists(ultimate_end_user)
+        ultimate_end_user.delete()
+
+        return JsonResponse(data={'ultimate_end_user': 'deleted'}, status=status.HTTP_200_OK)
 
 
 class ApplicationConsignee(APIView):
     authentication_classes = (ExporterAuthentication,)
 
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=True)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def post(self, request, application):
         """
         Create a consignee and add it to a application
         """
         data = request.data
-        data['organisation'] = str(request.user.organisation.id)
+        data['organisation'] = request.user.organisation.id
 
         serializer = ConsigneeSerializer(data=data)
         if serializer.is_valid():
@@ -127,17 +138,33 @@ class ApplicationConsignee(APIView):
                 delete_party_document_if_exists(previous_consignee)
                 previous_consignee.delete()
 
-            return JsonResponse(data={'consignee': serializer.data},
-                                status=status.HTTP_201_CREATED)
+            return JsonResponse(data={'consignee': serializer.data}, status=status.HTTP_201_CREATED)
 
-        return JsonResponse(data={'errors': serializer.errors},
-                            status=status.HTTP_400_BAD_REQUEST)
+        return JsonResponse(data={'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
+    @authorised_users(ExporterUser)
+    def delete(self, request, application):
+        """
+        Delete a consignee and their document from an application
+        """
+        consignee = application.consignee
+
+        if not consignee:
+            return JsonResponse(data={'errors': 'consignee not found'}, status=status.HTTP_404_NOT_FOUND)
+
+        application.consignee = None
+        application.save()
+        delete_party_document_if_exists(consignee)
+        consignee.delete()
+
+        return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 
 
 class ApplicationThirdParties(APIView):
     authentication_classes = (ExporterAuthentication,)
 
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=False)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def get(self, request, application):
         """
@@ -147,75 +174,41 @@ class ApplicationThirdParties(APIView):
 
         return JsonResponse(data={'third_parties': third_party_data})
 
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=True)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def post(self, request, application):
         """
         Create a third party and add it to a application
         """
         data = request.data
-        data['organisation'] = str(request.user.organisation.id)
+        data['organisation'] = request.user.organisation.id
 
         serializer = ThirdPartySerializer(data=data)
         if serializer.is_valid():
             third_party = serializer.save()
-            application.third_parties.add(str(third_party.id))
-            application.save()
+            application.third_parties.add(third_party.id)
 
-            return JsonResponse(data={'third_party': serializer.data},
-                                status=status.HTTP_201_CREATED)
+            return JsonResponse(data={'third_party': serializer.data}, status=status.HTTP_201_CREATED)
 
-        return JsonResponse(data={'errors': serializer.errors},
-                            status=status.HTTP_400_BAD_REQUEST)
-
-
-class RemoveApplicationUltimateEndUser(APIView):
-    authentication_classes = (ExporterAuthentication,)
-
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=True)
-    @authorised_users(ExporterUser)
-    def delete(self, request, application, ueu_pk):
-        """
-        Delete an ultimate end user and remove it from the application
-        """
-        try:
-            ultimate_end_user = UltimateEndUser.objects.get(id=ueu_pk)
-        except UltimateEndUser.DoesNotExist:
-            return JsonResponse(data={'errors': 'request invalid'},
-                                status=400)
-
-        if ultimate_end_user.organisation != request.user.organisation:
-            return JsonResponse(data={'errors': 'request invalid'},
-                                status=400)
-
-        application.ultimate_end_users.remove(str(ultimate_end_user.id))
-        application.save()
-        ultimate_end_user.delete()
-
-        return JsonResponse(data={'ultimate_end_user': 'deleted'})
+        return JsonResponse(data={'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class RemoveThirdParty(APIView):
     authentication_classes = (ExporterAuthentication,)
 
-    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, can_be_edited=True)
+    @only_applications(ApplicationLicenceType.STANDARD_LICENCE, in_a_major_edit_state=False)
     @authorised_users(ExporterUser)
     def delete(self, request, application, tp_pk):
         """
         Delete a third party and remove it from the application
         """
         try:
-            third_party = ThirdParty.objects.get(pk=tp_pk)
+            third_party = application.third_parties.get(pk=tp_pk)
         except ThirdParty.DoesNotExist:
-            return JsonResponse(data={'errors': 'request invalid'},
-                                status=400)
+            return JsonResponse(data={'errors': 'third party not found'}, status=status.HTTP_404_NOT_FOUND)
 
-        if third_party.organisation != request.user.organisation:
-            return JsonResponse(data={'errors': 'request invalid'},
-                                status=400)
-
-        application.third_parties.remove(str(third_party.id))
-        application.save()
+        application.third_parties.remove(third_party.id)
+        delete_party_document_if_exists(third_party)
         third_party.delete()
 
-        return JsonResponse(data={'third_party': 'deleted'})
+        return JsonResponse(data={'third_party': 'deleted'}, status=status.HTTP_200_OK)
